drastically reduce memory usage of chunks

+++ data/src/biomes.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-pub fn get_biome_ids() -> HashMap<String, i32> {
+pub fn get_biome_ids() -> HashMap<String, u8> {
 	return vec![
   ("minecraft:badlands", 0),
   ("minecraft:bamboo_jungle", 1),
